refactor: replies controller
